add coroutine

+++ app/src/main/java/com/hidero/test/ui/activities/SplashActivity.kt
@@ -3,13 +3,11 @@ package com.hidero.test.ui.activities
 import android.content.Intent
 import android.os.Bundle
 import android.os.Handler
-import android.view.animation.AnimationUtils
 import androidx.appcompat.app.AppCompatActivity
 import com.daimajia.androidanimations.library.Techniques
 import com.daimajia.androidanimations.library.YoYo
 import com.hidero.test.R
 import com.hidero.test.util.SPLASH_TIME_OUT
-import kotlinx.android.synthetic.main.activity_splash.*
 
 
 class SplashActivity : AppCompatActivity() {
@@ -29,7 +27,7 @@ class SplashActivity : AppCompatActivity() {
             startActivity(Intent(this@SplashActivity, MainActivity::class.java))
             finish()
             overridePendingTransition(R.anim.slide_in_left, R.anim.slide_fade_out)
-        }, SPLASH_TIME_OUT.toLong())
+        }, SPLASH_TIME_OUT)
     }
 
 }
